Move the instantiation part of dslash.h to instantiate_dslash.h, and let mdw_fused_dslash and dslash5_mobius_eofa to include only instantiate_dslash.h. This makes it possible to move these two out of QUDA_DSLASH_OBJS.
